Update the managed-ClickHouse page for specifics on stored fields (#661)
Fixes DOC-353

Updates the managed ClickHouse page for specific on the stored run and
feedback data. Added a table to define the run fields and whether
they're stored in obj storage or CH. Added the feedback field table to a
snippet and included on this page and the Feedback Data Format page.

## Preview


https://langchain-5e9cc07a-preview-clarif-1758897952-a148520.mintlify.app/langsmith/langsmith-managed-clickhouse