mark agent auth feature as beta (#730)
we discussed two days ago with marketing and GTM that the agent auth
feature will be a Beta as of the end of month launch week, so this PR
marks it as such

i have no strong opinions on the best way to show this in our docs, or
if it needs to be more extensive. feel free to suggest any other way of
showing this.

---------

Co-authored-by: Kathryn May <kathryn@langchain.dev>
Co-authored-by: Lauren Hirata Singh <lauren@langchain.dev>