Update the run evals api only content for endpoint (#828)
Moved the content from the old repository to the new Mintlify site for
this page. It looks like something unintentional happened during the
migration and the API endpoints in the example were messed up.

This uses the old content, only changes were to adding an OpenAI call,
which seemed to be causing issues due to versioning. Otherwise, this is
the same content from the old repository.

## Preview 


https://langchain-5e9cc07a-preview-evalua-1760126538-0eadc5b.mintlify.app/langsmith/run-evals-api-only