Improve double texting overview for confusing support text (#973)
On the double texting overview page, there was a sentence about it not
being included in LG OSS, however the phrasing made it seem like users
should use the described options in order to get around that. This PR
corrects that and improves some of the content on the page generally.