feat(langchain): `langchain-azure-storage` loader documentation (#1080)
Added the `AzureBlobStorage` page and removed the previous
`AzureBlobStorageFile` and `AzureBlobStorageContainer` pages as well as
any links to them. Added redirects for these pages as well, so now they
route to the `AzureBlobStorage` page, which contains a message with
links to the previous API reference if someone were to need them.

## Checklist
<!-- Put an 'x' in all boxes that apply -->
- [x] I have read the [contributing guidelines](README.md)
- [x] I have tested my changes locally using `docs dev`
- [x] All code examples have been tested and work correctly
- [x] I have used **root relative** paths for internal links
- [x] I have updated navigation in `src/docs.json` if needed
- I have gotten approval from the relevant reviewers

---------

Co-authored-by: Mason Daugherty <mason@langchain.dev>
Co-authored-by: Mason Daugherty <github@mdrxy.com>