feat: add `langchain-azure-storage` to reference docs (#1040)
Added
[langchain-azure-storage](https://pypi.org/project/langchain-azure-storage/)
to the reference docs with a new documentation page. We also plan to
create another PR to update the main LangChain documentation so that it
points to our newly created document loader instead of a previous
implementation. This PR is only to update the reference docs.

## Checklist
<!-- Put an 'x' in all boxes that apply -->
- [x] I have read the [contributing guidelines](README.md)
- [x] I have tested my changes locally using `docs dev`
- [x] All code examples have been tested and work correctly
- [x] I have used **root relative** paths for internal links
- [ ] I have updated navigation in `src/docs.json` if needed
- I have gotten approval from the relevant reviewers
- (Internal team members only / optional) I have created a preview
deployment using the [Create Preview Branch
workflow](https://github.com/langchain-ai/docs/actions/workflows/create-preview-branch.yml)

---------

Co-authored-by: Mason Daugherty <github@mdrxy.com>